+++ README.md
@@ -24,8 +24,8 @@ npm install @brightspace-ui/intl
 Then `import` only the functionality you need:
 
 ```javascript
-import {formatDate, formatTime} from '@brightspace-ui/intl/src/dateTime.js';
-import {formatNumber, formatPercent} from '@brightspace-ui/intl/src/number.js';
+import {formatDate, formatTime} from '@brightspace-ui/intl/lib/dateTime.js';
+import {formatNumber, formatPercent} from '@brightspace-ui/intl/lib/number.js';
 ```
 
 All of the APIs will automatically detect the document's language via the `lang` attribute on the `<html>` element. They'll also look for various `data-` attributes that will be present in Brightspace pages to access override and timezone information.
@@ -35,7 +35,7 @@ All of the APIs will automatically detect the document's language via the `lang`
 Integer and decimal numbers can be formatted in the user's locale using `formatNumber`. Percentages can be formatted using `formatPercent`. Use the optional `options` parameter for rounding.
 
 ```javascript
-import {formatNumber, formatPercent} from '@brightspace-ui/intl/src/number.js';
+import {formatNumber, formatPercent} from '@brightspace-ui/intl/lib/number.js';
 
 const number = formatNumber(8902.72, [options]); // -> '8,902.72' in en-US
 const percent = formatPercent(0.333, [options]); // -> '33.3 %' in en-US
@@ -48,7 +48,7 @@ Options:
 Formatting as an integer (rounded to 0 decimal places):
 
 ```javascript
-import {formatNumber} from '@brightspace-ui/intl/src/number.js';
+import {formatNumber} from '@brightspace-ui/intl/lib/number.js';
 
 const value = formatNumber(89.72, {
 	maximumFractionDigits: 0
@@ -58,7 +58,7 @@ const value = formatNumber(89.72, {
 Formatting as a percentage (rounded to 2 decimal places, but always showing at least 2 decimals):
 
 ```javascript
-import {formatPercent} from '@brightspace-ui/intl/src/number.js';
+import {formatPercent} from '@brightspace-ui/intl/lib/number.js';
 
 const value = formatPercent(0.333, {
 	minimumFractionDigits: 2,
@@ -71,7 +71,7 @@ const value = formatPercent(0.333, {
 The `parseNumber` method can be used to parse an integer or decimal number written in the user's locale.
 
 ```javascript
-import {parseNumber} from '@brightspace-ui/intl/src/number.js';
+import {parseNumber} from '@brightspace-ui/intl/lib/number.js';
 
 const value = parseNumber('-8 942,39'); // -> -8942.39 in fr-CA
 ```
@@ -83,7 +83,7 @@ Dates and times can be formatted in the user's locale using `formatDate`, `forma
 Combined dates and times are formatted using `formatDateTime`:
 
 ```javascript
-import {formatDateTime} from '@brightspace-ui/intl/src/dateTime.js';
+import {formatDateTime} from '@brightspace-ui/intl/lib/dateTime.js';
 
 const date = formatDateTime(
 	new Date(2015, 8, 23, 14, 5),
@@ -100,7 +100,7 @@ Options:
 To format a **date only** (without the time portion), use `formatDate`:
 
 ```javascript
-import {formatDate} from '@brightspace-ui/intl/src/dateTime.js';
+import {formatDate} from '@brightspace-ui/intl/lib/dateTime.js';
 
 const value = formatDate(
 	ew Date(2015, 8, 23),
@@ -123,7 +123,7 @@ Options:
 To format a **time only** (without the date portion), use `formatTime`:
 
 ```javascript
-import {formatTime} from '@brightspace-ui/intl/src/dateTime.js';
+import {formatTime} from '@brightspace-ui/intl/lib/dateTime.js';
 
 const time = formatTime(
 	new Date(2015, 8, 23, 14, 5)
@@ -140,7 +140,7 @@ Options:
 To parse a date written in the user's locale, use `parseDate`:
 
 ```javascript
-import {parseDate} from '@brightspace-ui/intl/src/dateTime.js';
+import {parseDate} from '@brightspace-ui/intl/lib/dateTime.js';
 
 const date = parseDate('2015-09-23'); // in fr-CA
 date.getFullYear(); // -> 2015
@@ -153,7 +153,7 @@ date.getDate(); // -> 23
 To parse a time written in the user's locale, use `parseTime`:
 
 ```javascript
-import {parseTime} from '@brightspace-ui/intl/src/dateTime.js';
+import {parseTime} from '@brightspace-ui/intl/lib/dateTime.js';
 
 const date = parseTime('14 h 05'); // in fr-CA
 date.getHours(); // -> 14
@@ -165,7 +165,7 @@ date.getMinutes(); // -> 5
 Use `formatFileSize` to format a file size appropriately for the user's locale.
 
 ```javascript
-import {formatFileSize} from '@brightspace-ui/intl/src/fileSize.js';
+import {formatFileSize} from '@brightspace-ui/intl/lib/fileSize.js';
 
 const fileSize = formatFileSize(100); // -> '100 bytes' in en-US
 ```
