+++ lib/common.js
