+++ lib/documentSettings.js
