+++ lib/fileSize.js
