+++ lib/number.js
