+++ package.json
@@ -2,20 +2,20 @@
   "name": "@brightspace-ui/intl",
   "version": "3.0.0",
   "description": "D2L internationalization APIs for number, date and time formatting and parsing.",
-  "main": "src/index.js",
-  "types": "src/index.d.ts",
-  "module": "src/index.js",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "module": "lib/index.js",
   "scripts": {
-    "lint": "eslint src test --ext .js",
+    "lint": "eslint lib test --ext .js",
     "test:unit": "concurrently -p name -n serve,test -s first -k \"http-server -p 8080 . -s\" \"mocha-headless-chrome -f http://localhost:8080/test/index.html -a no-sandbox -a disable-setuid-sandbox\"",
     "test": "npm run lint -s && npm run test:unit -s"
   },
   "files": [
-    "/src"
+    "/lib"
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/Brightspace/intl.git"
+    "url": "https://github.com/BrightspaceUI/intl.git"
   },
   "keywords": [
     "d2l",
@@ -30,9 +30,9 @@
   "author": "D2L Corporation",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/Brightspace/intl/issues"
+    "url": "https://github.com/BrightspaceUI/intl/issues"
   },
-  "homepage": "https://github.com/Brightspace/intl",
+  "homepage": "https://github.com/BrightspaceUI/intl",
   "devDependencies": {
     "chai": "^4",
     "concurrently": "^5",
